fix: relocate data
