+++ src/views/lists/addTaskToStore.ts
@@ -1,6 +1,6 @@
 import addTask from "../../actions/addTask";
 import appStore from "../../appStore";
-import { data } from "../../core/data";
+import { data } from "../../constants/data";
 
 appStore.subscribe(() => {
   console.log(appStore.getState());
